+++ jsbml.log
@@ -1,0 +1,5 @@
+2021-11-09 12:46:36,753 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-09 12:48:56,831 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-09 12:50:07,782 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-09 13:04:11,877 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
+2021-11-09 13:06:20,341 WARN o.j.SEDMLReader [main] Unexpected [Element: <annotation [Namespace: http://sed-ml.org/sed-ml/level1/version2]/>]
